Split ReminderModal into SystemNotifier, NotificationModal, and ReminderNotifier
ReminderModal mixed four responsibilities: choosing which surface(s) to
show a reminder on, the popup modal itself, and the full lifecycle of
Electron system notifications (tracking, replacing, and closing them).
Split these apart:

- system-notification.ts: SystemNotifier, which owns all Electron
  dependence (window.require("electron")) and the tracked-notification
  map; isSystemNotification()'s mobile check becomes
  SystemNotifier.isAvailable().
- reminder-modal.ts: NotificationModal, moved unchanged.
- reminder-notifier.ts (renamed from reminder.ts): ReminderNotifier
  (renamed from ReminderModal), which now only orchestrates which
  surface(s) to use and delegates sync()/destroy() to the toast manager
  and SystemNotifier.

Updates the reminderModal field/import in index.ts, and stale
`ReminderModal` references in comments in dnd-status-bar.ts and
reminder-toast.ts, to match. Pure structural refactor; behavior is
unchanged.
